README.md
This folder conatins the files for Task 1 - Prediction Using Supervised Machine Learning

In this task, we have to develop a model to predict the percentage of marks a student is expected to score based on the number of hours he/she studied. This is implemented using Simple Linear Regression involving 2 variables.
This folder conatins the files for Task 1 - Prediction Using Supervised Machine Learning

https://www.kaggle.com/abs7781/tsf-task1-prediction-using-supervised-ml